% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_bar}
\alias{progress_bar}
\title{progress_bar}
\usage{
progress_bar(type, con, session, id, title = NULL, status = NULL,
  range_value = NULL, unit_mark = "\%")
}
\arguments{
\item{type}{(from httr::progress doc) Type of progress to display: either number of bytes uploaded or downloaded.}

\item{con}{(from httr::progress doc) Connection to send output too. Usually \code{stdout()} or \code{stderr}.}

\item{session}{(from shinyWidgets::updateProgressBar doc) The 'session' object passed to function given to shinyServer.}

\item{id}{(from shinyWidgets::updateProgressBar doc) An id used to update the progress bar.}

\item{title}{(from shinyWidgets::updateProgressBar doc) character, optional title.}

\item{status}{(from shinyWidgets::updateProgressBar doc) Color, must be a valid Bootstrap status : primary, info, success, warning, danger.}

\item{range_value}{(from shinyWidgets::updateProgressBar doc) Default is to display percentage ([0, 100]), but you can specify a custom range, e.g. -50, 50.}

\item{unit_mark}{(from shinyWidgets::updateProgressBar doc) Unit for value displayed on the progress bar, default to "\%".}
}
\value{
a function with rules to print out the progress.
}
\description{
same as httr:::progress_bar but with capability to talk to shinyWidgets::progressBar.
}
\seealso{
\code{\link{progress}}
}
