% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample(display.mode = "showcase")
}
\arguments{
\item{display.mode}{The mode in which to display the example. Defaults to showcase, but may be set to normal to see the example without code or commentary.}
}
\description{
Launch shiny example application using shinyhttr::progress_bar. This app also uses module to show that it works with it too.
}
